requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
bandit==1.8.3
beautifulsoup4==4.13.4
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.2
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
filelock==3.18.0
flake8==7.2.0
identify==2.6.12
idna==3.10
intasend-python==1.1.2
isort==6.0.1
kombu==5.5.4
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
paypalrestsdk==1.13.3
pbr==6.1.1
pdf2image==1.17.0
pillow==11.2.1
platformdirs==4.3.8
pre_commit==4.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.1.0
pypdf==5.6.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.5.0
PyYAML==6.0.2
redis==6.2.0
reportlab==4.4.2
requests==2.32.3
rich==14.0.0
rsa==4.9.1
sentry-sdk==2.34.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stevedore==5.4.1
stripe==12.1.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
python-dotenv==1.2.1
